feat(bind_models): add Python bindings for speech models
- Added new `bind_models` module with Python bindings for `Denoiser`, `SileroVad`, and `SpeechTimestamp`
- Implemented full constructor and methods for `SileroVad` and `SpeechTimestamp`
- Fixed integer division bug in `timestamp_t::start_s()` and `timestamp_t::end_s()`
- Added tests for Python bindings in `test_models.py`
- Verified end-to-end functionality with 37 DSP tests and 12 Python tests

+++ include/libspeech/models/silero_vad.h
@@ -33,11 +33,11 @@ class timestamp_t {
     }
 
     float start_s() const {
-        return start / sample_rate;
+        return static_cast<float>(start) / static_cast<float>(sample_rate);
     }
 
     float end_s() const {
-        return end / sample_rate;
+        return static_cast<float>(end) / static_cast<float>(sample_rate);
     }
 
    public:
